test(config): move platform-neutral config tests into config module
